requirements.txt
flask>=2.0.0
requests>=2.25.0
anthropic>=0.16.0
pandas>=2.0.0
python-dotenv>=1.0.0